% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthinto.R
\name{orthinto}
\alias{orthinto}
\title{Numerical Integration for Exploring Orthogonal Functions}
\usage{
orthinto(
  phi = function(x) cos(x),
  psi = function(x) sin(x),
  dl = seq(-pi, 0, pi/10),
  du = seq(pi, 0, -pi/10),
  ...
)
}
\arguments{
\item{phi}{A function of one variable (default: cosine function)}

\item{psi}{A function of one variable (default: sine function)}

\item{dl}{Numeric vector of lower integration bounds (default: seq(-pi, 0, pi/10))}

\item{du}{Numeric vector of upper integration bounds (default: seq(pi, 0, -pi/10))}

\item{...}{Additional arguments passed to \code{\link[stats]{integrate}}}
}
\value{
An object of class "oio" (Orthogonal Integrals are 0) containing:
\describe{
\item{oio}{A list of integration results from \code{\link[stats]{integrate}}}
\item{dl}{Vector of lower bounds used}
\item{du}{Vector of upper bounds used}
\item{phi_name}{Name of the phi function}
\item{psi_name}{Name of the psi function}
}
}
\description{
This function computes the integral of the product of two functions phi and psi
over specified intervals with the goal of exploring orthogonal functions. Two
functions are orthogonal over an interval if the integral of their product equals
zero over that interval. This is based on the definition from Chapter 1, Section 5
of Fourier Series by Georgi P. Tolstov.
}
\details{
The function computes integrals of the form:
\deqn{\int_{dl_i}^{du_i} \phi(x) \psi(x) dx}

If dl and du have different lengths, they are recycled to match the longer vector.
The default functions (cosine and sine) are orthogonal over the interval -pi to pi.
}
\examples{
# Test orthogonality of cosine and sine (should be near zero)
result1 <- orthinto()
print(result1)

# Test with custom functions
result2 <- orthinto(
  phi = function(x) x,
  psi = function(x) x^2,
  dl = -1,
  du = 1
)
summary(result2)

}
\references{
Tolstov, G. P. (1962). Fourier Series. Translated by R. A. Silverman. Dover Publications, Inc., New York
}
