% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthinto.R
\name{plot.oio}
\alias{plot.oio}
\title{Plot Method for oio Objects}
\usage{
\method{plot}{oio}(x, type = "intervals", ...)
}
\arguments{
\item{x}{An object of class "oio"}

\item{type}{Character string specifying plot type: "intervals" (default), "values", or "errors"}

\item{...}{Additional arguments passed to plot functions}
}
\description{
Plot Method for oio Objects
}
