% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthinto.R
\name{print.oio}
\alias{print.oio}
\title{Print Method for oio Objects}
\usage{
\method{print}{oio}(x, ...)
}
\arguments{
\item{x}{An object of class "oio"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print Method for oio Objects
}
