% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthinto.R
\name{summary.oio}
\alias{summary.oio}
\title{Summary Method for oio Objects}
\usage{
\method{summary}{oio}(object, ...)
}
\arguments{
\item{object}{An object of class "oio"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary Method for oio Objects
}
